feat: improve code quality
